Create additional security logic in order and product command/query handlers

+++ StyleShop.Application/Order/Commands/DeleteOrder/DeleteOrderCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MediatR;
+using StyleShop.Application.ApplicationUser;
 using StyleShop.Domain.Interfaces;
 
 namespace StyleShop.Application.Order.Commands.DeleteOrder
@@ -10,14 +11,24 @@ namespace StyleShop.Application.Order.Commands.DeleteOrder
 
         private readonly IMapper _mapper;
 
-        public DeleteOrderCommandHandler(IOrderRepository orderRepository, IMapper mapper)
+        private readonly IUserContext _userContext;
+
+        public DeleteOrderCommandHandler(IOrderRepository orderRepository, IMapper mapper, IUserContext userContext)
         {
             _orderRepository = orderRepository;
             _mapper = mapper;
+            _userContext = userContext; 
         }
 
         public async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
         {
+            var currentUser = _userContext.GetCurrentUser();
+
+            if (!currentUser.IsInRole("admin"))
+            {
+                return Unit.Value;
+            }
+
             var order = _mapper.Map<Domain.Entities.Order>(request);
 
             _orderRepository.Delete(order);
