+++ StyleShop.Application/Product/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MediatR;
+using StyleShop.Application.ApplicationUser;
 using StyleShop.Domain.Interfaces;
 
 namespace StyleShop.Application.Product.Commands.DeleteProduct
@@ -10,14 +11,24 @@ namespace StyleShop.Application.Product.Commands.DeleteProduct
 
         private readonly IMapper _mapper;   
 
-        public DeleteProductCommandHandler(IProductRepository productRepository, IMapper mapper)
+        private readonly IUserContext _userContext; 
+
+        public DeleteProductCommandHandler(IProductRepository productRepository, IMapper mapper, IUserContext userContext)
         {
             _productRepository = productRepository;
             _mapper = mapper;
+            _userContext = userContext;
         }
 
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
+            var currentUser = _userContext.GetCurrentUser();
+
+            if (!currentUser.IsInRole("admin"))
+            {
+                return Unit.Value;
+            }
+
             var product = _mapper.Map<Domain.Entities.Product>(request);
            
             _productRepository.Delete(product);
